[refactor]: 사용자 인증 로직을 User -> Auth Domain으로 변경
