+++ backend/src/main/java/com/ddbb/dingdong/domain/user/service/UserErrors.java
@@ -4,7 +4,6 @@ import com.ddbb.dingdong.domain.common.exception.ErrorInfo;
 
 public enum UserErrors implements ErrorInfo {
     NOT_FOUND("해당 유저를 찾을 수 없습니다"),
-    NOT_MATCHED_PASSWORD("패스워드가 일치하지 않습니다"),
     ;
 
     private final String desc;
